+++ blocklets/ai-kit/api/src/libs/ai-provider.ts
@@ -1,3 +1,4 @@
+import { HttpsProxyAgent } from 'https-proxy-agent';
 import { OpenAI } from 'openai';
 
 import env from './env';
@@ -5,11 +6,14 @@ import env from './env';
 let currentApiKeyIndex = 0;
 
 export function getAIProvider() {
-  const { openaiApiKey } = env;
+  const { openaiApiKey, httpsProxy } = env;
 
   const apiKey = openaiApiKey[currentApiKeyIndex++ % openaiApiKey.length];
 
   if (!apiKey) throw new Error('Missing required openai apiKey');
 
-  return new OpenAI({ apiKey });
+  return new OpenAI({
+    apiKey,
+    httpAgent: httpsProxy ? new HttpsProxyAgent(httpsProxy) : undefined,
+  });
 }