+++ packages/react/src/components/conversation/conversation.tsx
@@ -1,7 +1,7 @@
 import { Error } from '@mui/icons-material';
 import { Alert, Avatar, Box, BoxProps, CircularProgress } from '@mui/material';
 import isNil from 'lodash/isNil';
-import { ChatCompletionRequestMessage } from 'openai';
+import { ChatCompletionMessageParam } from 'openai/resources';
 import { ReactNode, RefObject, forwardRef, useCallback, useEffect, useImperativeHandle, useRef } from 'react';
 
 import ImagePreview from '../image-preview';
@@ -10,7 +10,7 @@ import Prompt, { PromptProps } from './prompt';
 
 export interface MessageItem {
   id: string;
-  prompt?: string | ChatCompletionRequestMessage[];
+  prompt?: string | ChatCompletionMessageParam[];
   response?: string | { url: string }[];
   loading?: boolean;
   error?: { message: string };