+++ blocklets/ai-kit/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.80 (2023-11-14)
+
+- chore: set proxy url for ai
+- chore: upgrade openai
+
 ## 0.0.79 (2023-11-13)
 
 - feat: support dall-e-3